Move on_transform_command into parser.
Which involved as surprising amount of test refactorings.
